+++ js/listener.js
@@ -37,13 +37,22 @@
     };
 
     OCA.Onlyoffice.onRequestSaveAs = function (saveData) {
+
+        var arrayPath = OCA.Viewer.file.split("/");
+        arrayPath.pop();
+        arrayPath.shift();
+        var currentDir = "/" + arrayPath.join("/");
+
         OC.dialogs.filepicker(t(OCA.Onlyoffice.AppName, "Save as"),
             function (fileDir) {
                 saveData.dir = fileDir;
                 $(OCA.Onlyoffice.frameSelector)[0].contentWindow.OCA.Onlyoffice.editorSaveAs(saveData);
             },
             false,
-            "httpd/unix-directory");
+            "httpd/unix-directory",
+            false,
+            OC.dialogs.FILEPICKER_TYPE_CHOOSE,
+            currentDir);
     };
 
     OCA.Onlyoffice.onRequestInsertImage = function (imageMimes) {