+++ js/main.js
@@ -140,6 +140,8 @@
     };
 
     OCA.Onlyoffice.CloseEditor = function () {
+        OCA.Onlyoffice.frameSelector = null;
+
         $("body").removeClass("onlyoffice-inline");
         $("#onlyofficeHeader").remove();
 
@@ -381,6 +383,6 @@
         }
     };
 
-    $(document).ready(initPage);
+    initPage();
 
 })(OCA);