+++ js/viewer.js
@@ -91,7 +91,9 @@
             OCA.Onlyoffice.GetSettings(function () {
 
                 var mimes = $.map(OCA.Onlyoffice.setting.formats, function (format) {
-                    return format.mime;
+                    if (format.def) {
+                        return format.mime;
+                    }
                 });
 
                 OCA.Viewer.registerHandler({
@@ -104,6 +106,6 @@
         }
     };
 
-    $(document).ready(initPage)
+    initPage();
 
 })(OCA);