+++ lib/appconfig.php
@@ -215,6 +215,13 @@ class AppConfig {
     const WATERMARK_APP_NAMESPACE = "files";
 
     /**
+     * The config key for limit thumbnail size
+     *
+     * @var string
+     */
+    public $_limitThumbSize = "limit_thumb_size";
+
+    /**
      * The config key for the modifyFilter
      *
      * @var string
@@ -987,6 +994,21 @@ class AppConfig {
         }
 
         return $this->GetSystemValue($this->_verification);
+    }
+
+    /**
+     * Get the limit on size document when generating thumbnails
+     *
+     * @return int
+     */
+    public function GetLimitThumbSize() {
+        $limitSize = (integer)$this->GetSystemValue($this->_limitThumbSize);
+
+        if (!empty($limitSize)) {
+            return $limitSize;
+        }
+
+        return 100*1024*1024;
     }
 
     /**